+++ app/ts/DeterministicTab/DeterministicResults/SensitivityMeasurementsTable/SensitivityMeasurementsTableRow/SensitivityMeasurementsTableCell/SensitivityMeasurementsTableCell.tsx
@@ -1,12 +1,4 @@
-import {
-  Button,
-  Grid,
-  Popover,
-  Slider,
-  TableCell,
-  TextField
-} from '@material-ui/core';
-import {CSSProperties} from '@material-ui/core/styles/withStyles';
+import {Button, TableCell} from '@material-ui/core';
 import ICriterion from '@shared/interface/ICriterion';
 import {DeterministicResultsContext} from 'app/ts/DeterministicTab/DeterministicResultsContext/DeterministicResultsContext';
 import {
@@ -18,13 +10,8 @@ import significantDigits from 'app/ts/ManualInput/Util/significantDigits';
 import {SettingsContext} from 'app/ts/Settings/SettingsContext';
 import {SubproblemContext} from 'app/ts/Workspace/SubproblemContext/SubproblemContext';
 import {WorkspaceContext} from 'app/ts/Workspace/WorkspaceContext';
-import React, {
-  ChangeEvent,
-  MouseEvent,
-  useContext,
-  useEffect,
-  useState
-} from 'react';
+import React, {MouseEvent, useContext, useEffect, useState} from 'react';
+import SensitivityMeasurementsTablePopover from './SensitivityMeasurementsTablePopover/SensitivityMeasurementsTablePopover';
 
 export default function SensitivityMeasurementsTableCell({
   criterion,
@@ -51,7 +38,6 @@ export default function SensitivityMeasurementsTableCell({
   const [localValue, setLocalValue] = useState<number>(
     getPercentifiedValue(values.currentValue, usePercentage)
   );
-  const [inputError, setInputError] = useState<string>('');
 
   const min = getPercentifiedValue(
     currentSubproblem.definition.ranges[criterion.dataSources[0].id][0],
@@ -65,7 +51,6 @@ export default function SensitivityMeasurementsTableCell({
     getStepSizeForCriterion(criterion),
     usePercentage
   );
-  const marginTop: CSSProperties = {marginTop: '50px', textAlign: 'center'};
 
   useEffect(() => {
     if (isDirty && values.currentValue === values.originalValue) {
@@ -77,7 +62,7 @@ export default function SensitivityMeasurementsTableCell({
     setAnchorEl(event.currentTarget);
   }
 
-  function closePopover() {
+  function closePopover(inputError: string) {
     if (localValue === values.originalValue) {
       setIsDirty(false);
     } else {
@@ -90,26 +75,6 @@ export default function SensitivityMeasurementsTableCell({
       }
       setAnchorEl(null);
     }
-  }
-  function inputChanged(
-    event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ): void {
-    valueChanged(Number.parseFloat(event.target.value));
-  }
-
-  function sliderChanged(event: ChangeEvent<any>, newValue: number): void {
-    valueChanged(newValue);
-  }
-
-  function valueChanged(newValue: number) {
-    if (isNaN(newValue)) {
-      setInputError('Invalid value');
-    } else if (newValue < min || newValue > max) {
-      setInputError(`Value must be between ${min} and ${max}`);
-    } else {
-      setInputError('');
-    }
-    setLocalValue(newValue);
   }
 
   function getLabel(): string {
@@ -128,58 +93,15 @@ export default function SensitivityMeasurementsTableCell({
       <Button className="text-centered" onClick={openPopover} variant="text">
         <a> {getLabel()}</a>
       </Button>
-      <Popover
-        open={!!anchorEl}
-        onClose={closePopover}
-        anchorOrigin={{vertical: 'center', horizontal: 'center'}}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'center'
-        }}
+      <SensitivityMeasurementsTablePopover
         anchorEl={anchorEl}
-      >
-        <Grid container style={{minWidth: '300px', minHeight: '100px'}}>
-          <Grid item xs={2} style={marginTop}>
-            {min}
-          </Grid>
-          <Grid item xs={8} style={marginTop}>
-            <Slider
-              id="sensitivity-value-slider"
-              marks
-              valueLabelDisplay="on"
-              value={localValue}
-              min={min}
-              max={max}
-              onChange={sliderChanged}
-              step={stepSize}
-              track={false}
-            />
-          </Grid>
-          <Grid item xs={2} style={marginTop}>
-            {max}
-          </Grid>
-          <Grid
-            item
-            xs={12}
-            style={{textAlign: 'center', marginBottom: '20px'}}
-          >
-            <TextField
-              id="sensitivity-value-input"
-              value={localValue}
-              onChange={inputChanged}
-              type="number"
-              inputProps={{
-                min: min,
-                max: max,
-                step: stepSize
-              }}
-              error={!!inputError}
-              helperText={inputError ? inputError : ''}
-              autoFocus
-            />
-          </Grid>
-        </Grid>
-      </Popover>
+        closePopover={closePopover}
+        min={min}
+        max={max}
+        localValue={localValue}
+        setLocalValue={setLocalValue}
+        stepSize={stepSize}
+      />
     </TableCell>
   );
 }
